refactor(persistence): split crud.py into a per-domain crud/ package
Break the 2861-line crud.py monolith into ten domain submodules
(repository, pages, graph, external_systems, git, analysis, decisions,
chat, knowledge_graph) behind a star-import façade in crud/__init__.py.
Every previously-public name is still importable from
repowise.core.persistence.crud, so the ~68 dependents are untouched.

Collapse the five duplicated SELECT-then-INSERT/UPDATE loops (graph
nodes/edges/metrics, wiki symbols, git-metadata bulk) onto a single
generic _batch_upsert(model, items, key_fn, update_fn, insert_fn,
batch_size) in _shared.py, preserving the per-batch-flush semantics
(including no-flush-on-empty for the batched git path). Drop the
redundant `import json as _json` alias.

Add test_crud_smoke.py covering the gaps the existing test_crud.py
left: graph-metrics + git-bulk round trips (incl. the 500-row chunk
boundary) and the edge/symbol in-place update paths.

No behavior change. Full tests/unit green (2388 passed).
