+++ packages/core/src/repowise/core/persistence/crud/chat.py
@@ -1,0 +1,118 @@
+"""CRUD operations for the chat domain (repowise persistence layer).
+
+Split out of the former monolithic ``crud.py``; ``crud/__init__.py`` re-exports
+every public name, so existing imports are unaffected.
+"""
+
+from __future__ import annotations
+
+import json
+
+from sqlalchemy import func, select
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from ..models import (
+    ChatMessage,
+    Conversation,
+    _now_utc,
+)
+
+# ---------------------------------------------------------------------------
+# Conversation CRUD
+# ---------------------------------------------------------------------------
+
+
+async def create_conversation(
+    session: AsyncSession,
+    *,
+    repository_id: str,
+    title: str = "New conversation",
+) -> Conversation:
+    conv = Conversation(repository_id=repository_id, title=title)
+    session.add(conv)
+    await session.flush()
+    return conv
+
+
+async def get_conversation(session: AsyncSession, conversation_id: str) -> Conversation | None:
+    return await session.get(Conversation, conversation_id)
+
+
+async def list_conversations(
+    session: AsyncSession, repository_id: str, *, limit: int = 50
+) -> list[Conversation]:
+    result = await session.execute(
+        select(Conversation)
+        .where(Conversation.repository_id == repository_id)
+        .order_by(Conversation.updated_at.desc())
+        .limit(limit)
+    )
+    return list(result.scalars().all())
+
+
+async def update_conversation_title(
+    session: AsyncSession, conversation_id: str, title: str
+) -> Conversation | None:
+    conv = await session.get(Conversation, conversation_id)
+    if conv:
+        conv.title = title
+        conv.updated_at = _now_utc()
+        await session.flush()
+    return conv
+
+
+async def delete_conversation(session: AsyncSession, conversation_id: str) -> bool:
+    conv = await session.get(Conversation, conversation_id)
+    if conv is None:
+        return False
+    await session.delete(conv)
+    await session.flush()
+    return True
+
+
+async def touch_conversation(session: AsyncSession, conversation_id: str) -> None:
+    """Update the updated_at timestamp of a conversation."""
+    conv = await session.get(Conversation, conversation_id)
+    if conv:
+        conv.updated_at = _now_utc()
+        await session.flush()
+
+
+# ---------------------------------------------------------------------------
+# ChatMessage CRUD
+# ---------------------------------------------------------------------------
+
+
+async def create_chat_message(
+    session: AsyncSession,
+    *,
+    conversation_id: str,
+    role: str,
+    content: dict,
+) -> ChatMessage:
+    msg = ChatMessage(
+        conversation_id=conversation_id,
+        role=role,
+        content_json=json.dumps(content),
+    )
+    session.add(msg)
+    await session.flush()
+    return msg
+
+
+async def list_chat_messages(session: AsyncSession, conversation_id: str) -> list[ChatMessage]:
+    result = await session.execute(
+        select(ChatMessage)
+        .where(ChatMessage.conversation_id == conversation_id)
+        .order_by(ChatMessage.created_at.asc())
+    )
+    return list(result.scalars().all())
+
+
+async def count_chat_messages(session: AsyncSession, conversation_id: str) -> int:
+    result = await session.execute(
+        select(func.count())
+        .select_from(ChatMessage)
+        .where(ChatMessage.conversation_id == conversation_id)
+    )
+    return result.scalar() or 0
