+++ packages/core/src/repowise/core/persistence/crud/pages.py
@@ -1,0 +1,258 @@
+"""CRUD operations for the pages domain (repowise persistence layer).
+
+Split out of the former monolithic ``crud.py``; ``crud/__init__.py`` re-exports
+every public name, so existing imports are unaffected.
+"""
+
+from __future__ import annotations
+
+import json
+from datetime import datetime
+from typing import Any
+
+from sqlalchemy import select
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from ..models import (
+    Page,
+    PageVersion,
+    _new_uuid,
+    _now_utc,
+)
+from ._shared import _parse_dt
+
+# ---------------------------------------------------------------------------
+# Page CRUD (with versioning)
+# ---------------------------------------------------------------------------
+
+
+async def upsert_page(
+    session: AsyncSession,
+    *,
+    page_id: str,
+    repository_id: str,
+    page_type: str,
+    title: str,
+    content: str,
+    summary: str = "",
+    target_path: str,
+    source_hash: str,
+    model_name: str,
+    provider_name: str,
+    input_tokens: int = 0,
+    output_tokens: int = 0,
+    cached_tokens: int = 0,
+    generation_level: int = 0,
+    confidence: float = 1.0,
+    freshness_status: str = "fresh",
+    metadata: dict | None = None,
+    created_at: datetime | None = None,
+    updated_at: datetime | None = None,
+) -> Page:
+    """Insert or update a wiki page, creating a PageVersion snapshot on update.
+
+    First call  → inserts Page at version=1.
+    Subsequent  → archives the current Page as a PageVersion, then updates the
+                  Page in-place (version += 1, created_at preserved).
+    """
+    now = _now_utc()
+    page_created_at = created_at or now
+    page_updated_at = updated_at or now
+    meta_json = json.dumps(metadata or {})
+
+    existing_result = await session.execute(select(Page).where(Page.id == page_id))
+    existing = existing_result.scalar_one_or_none()
+
+    if existing is not None:
+        # Archive the current state before overwriting
+        snapshot = PageVersion(
+            id=_new_uuid(),
+            page_id=existing.id,
+            repository_id=existing.repository_id,
+            version=existing.version,
+            page_type=existing.page_type,
+            title=existing.title,
+            content=existing.content,
+            source_hash=existing.source_hash,
+            model_name=existing.model_name,
+            provider_name=existing.provider_name,
+            input_tokens=existing.input_tokens,
+            output_tokens=existing.output_tokens,
+            confidence=existing.confidence,
+            archived_at=now,
+        )
+        session.add(snapshot)
+
+        # Update Page in place (preserves created_at)
+        existing.page_type = page_type
+        existing.title = title
+        existing.content = content
+        existing.summary = summary
+        existing.target_path = target_path
+        existing.source_hash = source_hash
+        existing.model_name = model_name
+        existing.provider_name = provider_name
+        existing.input_tokens = input_tokens
+        existing.output_tokens = output_tokens
+        existing.cached_tokens = cached_tokens
+        existing.generation_level = generation_level
+        existing.version = existing.version + 1
+        existing.confidence = confidence
+        existing.freshness_status = freshness_status
+        existing.metadata_json = meta_json
+        existing.updated_at = page_updated_at
+
+        await session.flush()
+        return existing
+    else:
+        page = Page(
+            id=page_id,
+            repository_id=repository_id,
+            page_type=page_type,
+            title=title,
+            content=content,
+            summary=summary,
+            target_path=target_path,
+            source_hash=source_hash,
+            model_name=model_name,
+            provider_name=provider_name,
+            input_tokens=input_tokens,
+            output_tokens=output_tokens,
+            cached_tokens=cached_tokens,
+            generation_level=generation_level,
+            version=1,
+            confidence=confidence,
+            freshness_status=freshness_status,
+            metadata_json=meta_json,
+            created_at=page_created_at,
+            updated_at=page_updated_at,
+        )
+        session.add(page)
+        await session.flush()
+        return page
+
+
+async def load_prior_pages(
+    session: AsyncSession,
+    repository_id: str,
+) -> dict[str, Any]:
+    """Return a ``page_id → PriorPage`` map for cross-run cache reuse.
+
+    Loads every existing wiki page for the repository so the generator can
+    short-circuit the LLM call when the freshly rendered prompt produces a
+    matching ``source_hash`` under the same model. Returns an empty dict if
+    nothing has been generated yet.
+    """
+    # Import lazily — keeps persistence independent of generation models at
+    # module-load time.
+    from repowise.core.generation.page_generator import PriorPage
+
+    result = await session.execute(select(Page).where(Page.repository_id == repository_id))
+    prior: dict[str, Any] = {}
+    for row in result.scalars():
+        prior[row.id] = PriorPage(
+            source_hash=row.source_hash,
+            model_name=row.model_name,
+            content=row.content,
+            input_tokens=row.input_tokens,
+            output_tokens=row.output_tokens,
+            cached_tokens=row.cached_tokens,
+        )
+    return prior
+
+
+async def upsert_page_from_generated(
+    session: AsyncSession,
+    generated_page: object,  # repowise.core.generation.models.GeneratedPage
+    repository_id: str,
+) -> Page:
+    """Convenience wrapper that unpacks a GeneratedPage dataclass.
+
+    This keeps the CRUD layer independent of the generation models at the
+    import level while still providing a clean API for callers that have a
+    GeneratedPage in hand.
+    """
+    gp = generated_page  # type alias for brevity
+    return await upsert_page(
+        session,
+        page_id=gp.page_id,  # type: ignore[attr-defined]
+        repository_id=repository_id,
+        page_type=gp.page_type,  # type: ignore[attr-defined]
+        title=gp.title,  # type: ignore[attr-defined]
+        content=gp.content,  # type: ignore[attr-defined]
+        summary=getattr(gp, "summary", "") or "",
+        target_path=gp.target_path,  # type: ignore[attr-defined]
+        source_hash=gp.source_hash,  # type: ignore[attr-defined]
+        model_name=gp.model_name,  # type: ignore[attr-defined]
+        provider_name=gp.provider_name,  # type: ignore[attr-defined]
+        input_tokens=gp.input_tokens,  # type: ignore[attr-defined]
+        output_tokens=gp.output_tokens,  # type: ignore[attr-defined]
+        cached_tokens=gp.cached_tokens,  # type: ignore[attr-defined]
+        generation_level=gp.generation_level,  # type: ignore[attr-defined]
+        confidence=gp.confidence,  # type: ignore[attr-defined]
+        freshness_status=gp.freshness_status,  # type: ignore[attr-defined]
+        metadata=gp.metadata,  # type: ignore[attr-defined]
+        created_at=_parse_dt(gp.created_at),  # type: ignore[attr-defined]
+        updated_at=_parse_dt(gp.updated_at),  # type: ignore[attr-defined]
+    )
+
+
+async def get_page(session: AsyncSession, page_id: str) -> Page | None:
+    """Return a Page by its page_id, or None."""
+    return await session.get(Page, page_id)
+
+
+async def list_pages(
+    session: AsyncSession,
+    repository_id: str,
+    *,
+    page_type: str | None = None,
+    limit: int = 100,
+    offset: int = 0,
+    sort_by: str = "updated_at",
+    order: str = "desc",
+) -> list[Page]:
+    """Return pages for a repository, optionally filtered by page_type."""
+    q = select(Page).where(Page.repository_id == repository_id)
+    if page_type is not None:
+        q = q.where(Page.page_type == page_type)
+    _sort_cols = {
+        "updated_at": Page.updated_at,
+        "confidence": Page.confidence,
+        "created_at": Page.created_at,
+    }
+    sort_col = _sort_cols.get(sort_by, Page.updated_at)
+    q = q.order_by(sort_col.asc() if order == "asc" else sort_col.desc())
+    q = q.limit(limit).offset(offset)
+    result = await session.execute(q)
+    return list(result.scalars().all())
+
+
+async def get_page_versions(
+    session: AsyncSession,
+    page_id: str,
+    *,
+    limit: int = 50,
+) -> list[PageVersion]:
+    """Return historical versions of a page, newest first."""
+    result = await session.execute(
+        select(PageVersion)
+        .where(PageVersion.page_id == page_id)
+        .order_by(PageVersion.version.desc())
+        .limit(limit)
+    )
+    return list(result.scalars().all())
+
+
+async def get_stale_pages(
+    session: AsyncSession,
+    repository_id: str,
+) -> list[Page]:
+    """Return pages with freshness_status in ('stale', 'expired')."""
+    result = await session.execute(
+        select(Page).where(
+            Page.repository_id == repository_id,
+            Page.freshness_status.in_(["stale", "expired"]),
+        )
+    )
+    return list(result.scalars().all())
